requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.87.0
graphql-core==3.2.3
greenlet==2.0.1
h11==0.14.0
idna==3.4
pydantic==1.10.2
Pygments==2.13.0
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
starlette==0.21.0
strawberry-graphql==0.142.2
typing_extensions==4.4.0
uvicorn==0.19.0
